MOON-231: Add Controlled/Uncontrolled input
